requirements.txt
Django==3.2.12
djangorestframework==3.13.1
django-environ==0.8.1
django-phonenumber-field[phonenumbers]==6.1.0
